backend/fix/updated Interfaz Irecod credit and end epoind final, add apireponse

+++ Backend/h3-18-proptechback/h3-18-proptechback.CreditRecord/Models/Requets/DeudasRequest.cs
@@ -4,7 +4,7 @@ namespace h3_18_proptechback.CreditRecord.Models.Requets
 {
     public class DeudasRequest
     {
-        public int identificacion {  get; set; }
+        public long identificacion {  get; set; }
     }
 }
